+++ src/lib/index.d.ts
@@ -1,1 +1,2 @@
 declare module 'geojson-validation'
+declare module 'geolib'